feat(annotations): move annotations to stable (#823)
* refactor: move files and add annotationsapi to cognite client

* refactor: fix imports

* refactor: remove annotations from playground

* refactor: add annotation types to main types

* refactor: cleanup imports

* refactor: lint

* refactor: remove types file
