+++ packages/playground/src/types.ts
@@ -7,7 +7,6 @@ import {
 } from '@cognite/sdk';
 
 export * from './api/documents/types.gen';
-export * from './api/annotations/types';
 
 // This file is here mostly to allow apis to import { ... } from '../../types';
 // Overriding types should probably be done in their respective API endpoint files, where possible
