+++ packages/stable/src/api/annotations/annotationsApi.ts
@@ -12,7 +12,7 @@ import {
   AnnotationSuggest,
   AnnotationFilterRequest,
   AnnotationModel,
-} from './types';
+} from '../../types';
 
 export class AnnotationsAPI extends BaseResourceAPI<AnnotationModel> {
   /**
